LayoutTests/fast/files/blob-as-frame-url-expected.txt

Makes sure that using a Blob URL as frame URL works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframe.contentDocument.body.innerText is "FOO"
PASS iframe.contentWindow.location.href == blobURL is true
PASS iframe.contentDocument.URL == blobURL is true
PASS successfullyParsed is true

TEST COMPLETE
